+++ app/src/main/java/com/battlelancer/seriesguide/SgAppContainer.kt
@@ -6,6 +6,8 @@ package com.battlelancer.seriesguide
 import android.content.Context
 import com.battlelancer.seriesguide.diagnostics.DebugLogBuffer
 import com.battlelancer.seriesguide.util.PackageTools
+import com.battlelancer.seriesguide.util.PackageTools.isEuropeanEconomicArea
+import com.battlelancer.seriesguide.util.PackageTools.isUnitedStates
 import timber.log.Timber
 
 class SgAppContainer(context: Context) {
@@ -18,9 +20,20 @@ class SgAppContainer(context: Context) {
      */
     val preventExternalLinks by lazy {
         val installedByPlay = PackageTools.wasInstalledByPlayStore(context)
-        val isDeviceInEEA = PackageTools.isDeviceInEEA(context)
-        (installedByPlay && !isDeviceInEEA)
-            .also { Timber.d("preventExternalLinks = %s", it) }
+        val region = PackageTools.getDeviceRegion(context)
+        val isEEA = region.isEuropeanEconomicArea(context)
+        val isUS = region.isUnitedStates()
+        (installedByPlay && !isEEA && !isUS)
+            .also {
+                Timber.i(
+                    "preventExternalLinks=%s installedByPlay=%s region=%s isEEA=%s isUS=%s",
+                    installedByPlay,
+                    it,
+                    region.code,
+                    isEEA,
+                    isUS
+                )
+            }
 //            .let { if (BuildConfig.DEBUG) true else it }
     }